modif button for non admin

+++ site/functions.php
@@ -1,4 +1,5 @@
 <?php
+// session_start();
 function debug($variable){
     echo '<pre>' . print_r($variable, true) . '<pre>';
 }
@@ -7,11 +8,12 @@ function createPostButtons(int $id = -1){
     ?>
     <div class="crud">
         <a href="create.php" role="button" class="btn btn-outline-success ml-1">Add post</a>
-      <?php if($id >= 0){ ?>
+      <?php if($id >= 0 && $_SESSION['admin'] == true){ ?>
                 <a href="update.php?id=<?= $id ?>" role="button" class="btn btn-outline-warning ml-1">Edit post</a>
                 <a href="delete.php?id=<?= $id ?>" role="button" class="btn btn-outline-danger ml-1">Delete it</a>
-                <a href="posts.php" role="button" class="btn btn-outline-secondary ml-1">Go Back</a>
         <?php } ?>
+                <a href="posts.php" role="button" class="btn btn-outline-secondary ml-1">Go Back</a>
+
     </div>
     <?php
 }
